Redirect Handling Fixes #25 funzia

+++ src/ms_heredoc.c
@@ -6,7 +6,7 @@
 /*   By: marboccu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 20:32:32 by marboccu          #+#    #+#             */
-/*   Updated: 2024/06/10 11:33:04 by marboccu         ###   ########.fr       */
+/*   Updated: 2024/06/10 17:53:31 by marboccu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ static char *heredoc_read(t_var *mshell, t_list *token, int count)
 	return (heredoc_file);
 }
 
-void ms_heredoc(t_var *mshell, t_command *cmds)
+int ms_heredoc(t_var *mshell, t_command *cmds)
 {
 	t_list	*current;
 	char *heredoc_name;
@@ -67,12 +67,18 @@ void ms_heredoc(t_var *mshell, t_command *cmds)
 
 	count = 1;
 	heredoc_name = NULL;
+	filein_name = NULL;
 	current = cmds->in_redirects;
 	while (current != NULL)
 	{
 		if (str_equals(current->val, "<<"))
 		{
 			heredoc_name = heredoc_read(mshell, current, count++);
+			if (!heredoc_name)
+			{
+				cmds->in_file = NULL;
+				return (KO);
+			}
 			if (current->next->next != NULL)
 			{
 				unlink(heredoc_name);
@@ -80,68 +86,42 @@ void ms_heredoc(t_var *mshell, t_command *cmds)
 			}
 			current = current->next;
 		}
-		else if (str_equals(current->val, "<"))
+		current = current->next;
+	}
+	if (heredoc_name != NULL)
+	{
+		cmds->in_file = heredoc_name;
+	}
+	current = cmds->in_redirects;
+	while (current != NULL)
+	{
+		if (str_equals(current->val, "<"))
 		{
 			filein_name = current->next->val;
 			if (access(filein_name, F_OK) != 0)
 			{
-				ft_perror("minishell: %s No such file or directory\n", filein_name);
-				filein_name = ft_free(filein_name);
-				//printf("filein_name: %s\n", filein_name);
+				ft_perror("minishell: %s: No such file or directory\n", filein_name);
+				cmds->in_file = NULL;
+				return (KO);
 			}
-			else if (current->next->next == NULL)
+			if (current->next->next == NULL)
 			{
-				//cmds->in_file = filein_name;
+				//printf("filein_name: %s\n", filein_name);
+				cmds->in_file = str_dup(filein_name);
 				fd = open(filein_name, O_RDONLY);
 				if (fd == -1)
 				{
-					ft_perror("open");
+					pf_errcode(ERR_FILE_PERMISSION_DENIED);
 					cmds->in_file = NULL;
+					return (OK);
 				}
-				dup2(fd, STDIN_FILENO);
 				close(fd);
 			}
 			current = current->next;
-			printf("filein_nameee: %s\n", filein_name);
 		}
 		current = current->next;
 	}
-	printf("heredoc_name: %s\n", heredoc_name);
-	if (heredoc_name != NULL)
-		cmds->in_file = heredoc_name;
-	else if (filein_name != NULL)
-		cmds->in_file = filein_name;
-	else
-		cmds->in_file = NULL;
-	printf("cmds->in_file: %s\n", cmds->in_file);
-	// current = cmds->in_redirects;
-	// while (current != NULL)
-	// {
-	// 	if (str_equals(current->val, "<"))
-	// 	{
-	// 		filein_name = current->next->val;
-	// 		if (access(filein_name, F_OK) != 0)
-	// 		{
-	// 			ft_perror("file not found");
-	// 			cmds->in_file = NULL;
-	// 			return ;
-	// 		}
-	// 		else if (current->next->next == NULL)
-	// 		{
-	// 			cmds->in_file = filein_name;
-	// 			fd = open(filein_name, O_RDONLY);
-	// 			if (fd == -1)
-	// 			{
-	// 				ft_perror("open");
-	// 				cmds->in_file = NULL;
-	// 				return ;
-	// 			}
-	// 			dup2(fd, STDIN_FILENO);
-	// 			close(fd);
-	// 		}
-	// 		current = current->next;
-	// 	}
-	// 	current = current->next;
-	// }
+	return (OK);
+	//printf("cmds->in_file: %s\n", cmds->in_file);
 }
